+++ src/effects/kaleidoscope.js
@@ -1,39 +1,46 @@
 /**
  * @function kaleidoscope
  * @param {Object} [params]
- * @param {number} [params.segments=6] .
- * @param {number} [params.offset=0] .
+ * @param {number} [params.segments=6] number of times the view is divided.
+ * @param {number} [params.offset={x: 0.0, y: 0.0}] offset to move the source media from the center.
+ * @param {number} [params.rotation=0] extra angle to rotate the view.
  * @returns {kaleidoscopeEffect}
  *
  * @example kaleidoscope({segments: 12})
  */
-export default function ({ segments = 6, offset = 0 } = {}) {
+export default function ({ segments = 6, offset, rotation = 0 } = {}) {
+    const { x: offsetX, y: offsetY } = offset || { x: 0.0, y: 0.0 };
     /**
      * @typedef {Object} kaleidoscopeEffect
      * @property {number} segments
-     * @property {number} offset
+     * @property {{x: number?, y: number?}} offset
+     * @property {number} rotation
      * @property {boolean} disabled
      *
      * @example
      * effect.segments = 8;
-     * effect.offset = 0.5;
+     * effect.offset = { x: 0.5, y: 0.5 };
+     * effect.rotation = 45;
      */
     return {
         fragment: {
             uniform: {
                 u_kaleidoscopeEnabled: 'bool',
                 u_segments: 'float',
-                u_offset: 'float',
+                u_offset: 'vec2',
+                u_rotation: 'float',
             },
             source: `
-    if (u_kaleidoscopeEnabled) {
+    if (u_kaleidoscopeEnabled && u_segments > 0.0) {
         vec2 centered = v_texCoord - 0.5;
         float r = length(centered);
         float theta = atan(centered.y, centered.x);
-        theta = mod(theta, 2.0 * PI / u_segments);
+        theta = mod(theta, 2.0 * PI / u_segments) + radians(u_rotation);
         theta = abs(theta - PI / u_segments) - PI / u_segments;
         vec2 newCoords = r * vec2(cos(theta), sin(theta)) + 0.5;
-        sourceCoord = mod(newCoords - u_offset, 1.0);
+        sourceCoord = newCoords - u_offset;
+        // mirrored repeat
+        sourceCoord = mod(-sourceCoord, 1.0) * (mod(sourceCoord - 1.0, 2.0) - mod(sourceCoord, 1.0)) + mod(sourceCoord, 1.0) * (mod(sourceCoord, 2.0) - mod(sourceCoord, 1.0));
     }`,
         },
         get segments() {
@@ -43,10 +50,18 @@ export default function ({ segments = 6, offset = 0 } = {}) {
             this.uniforms[1].data[0] = +n;
         },
         get offset() {
-            return this.uniforms[2].data[0];
+            const [x, y] = this.uniforms[2].data;
+            return { x, y };
         },
-        set offset(o) {
-            this.uniforms[2].data[0] = +o;
+        set offset({ x, y }) {
+            if (typeof x !== 'undefined') this.uniforms[2].data[0] = x;
+            if (typeof y !== 'undefined') this.uniforms[2].data[1] = y;
+        },
+        get rotation() {
+            return this.uniforms[3].data[0];
+        },
+        set rotation(r) {
+            this.uniforms[3].data[0] = r;
         },
         get disabled() {
             return !this.uniforms[0].data[0];
@@ -68,7 +83,12 @@ export default function ({ segments = 6, offset = 0 } = {}) {
             {
                 name: 'u_offset',
                 type: 'f',
-                data: [offset],
+                data: [offsetX, offsetY],
+            },
+            {
+                name: 'u_rotation',
+                type: 'f',
+                data: [rotation],
             },
         ],
     };